refactor : #67 id 타입 전부 Long 타입으로 변경

+++ src/main/java/com/example/team1_be/domain/Apply/Apply.java
@@ -16,7 +16,7 @@ import javax.validation.constraints.NotNull;
 public class Apply {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Enumerated(EnumType.STRING)
     @NotNull
@@ -31,7 +31,7 @@ public class Apply {
     private Member member;
 
     @Builder
-    public Apply(int id, ApplyType state, Worktime worktime, Member member) {
+    public Apply(Long id, ApplyType state, Worktime worktime, Member member) {
         this.id = id;
         this.state = state;
         this.worktime = worktime;
