+++ src/main/java/com/example/team1_be/domain/Member/Member.java
@@ -16,7 +16,7 @@ import javax.validation.constraints.NotNull;
 public class Member {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(nullable = false)
     @NotNull
@@ -31,7 +31,7 @@ public class Member {
     private User user;
 
     @Builder
-    public Member(int id, Boolean isAdmin, Group group, User user) {
+    public Member(Long id, Boolean isAdmin, Group group, User user) {
         this.id = id;
         this.isAdmin = isAdmin;
         this.group = group;
